+++ docs/cli/configs/base.md
@@ -1,0 +1,1 @@
+::: prefect_dbt.cli.configs.base
