+++ docs/cli/configs/snowflake.md
@@ -1,0 +1,1 @@
+::: prefect_dbt.cli.configs.snowflake
