--- docs/cli/models.md
@@ -1,1 +1,0 @@
-::: prefect_dbt.cli.models
