+++ prefect_dbt/cli/configs/__init__.py
@@ -1,0 +1,6 @@
+from .base import TargetConfigs, GlobalConfigs  # noqa
+from .snowflake import (  # noqa
+    SnowflakeUserPasswordTargetConfigs,
+    SnowflakeKeyPairTargetConfigs,
+    SnowflakeSsoTargetConfigs,
+)
